+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - French translations by @anivon
 - @anivon localize Error parsing config message is JSON is invalid 
 - @blackadmin version typos in about window. 
+- Added support for selecting Warp and Ghostty as terminal backends.
 
 ## [1.2.9] - 2016-10-18
 ### Added